Update to bass v02r02 release.
x86 code is becoming somewhat presentable, with more and more addressing modes supported. Added a fun implicit sorting for addressing mode order. More powerful than the simple if(a.wildcard(b.pattern)) swap(a, b);
Started on some basic abstraction to eventually allow for toggling the default word size to 16-bits instead of 32-bits.
The next stumbling block is on opcode 0x63 - ARPL. It only accepts m16,r16 arguments. All of my routines always allow for both 16-bit and 32-bit, as everything else thus far has been m16/32 and r16/32. Will have to rework more to support it.
But yeah, x86 assembly is going to be slooooooooooooooooooow given the extreme complexity and ambiguity in the opcodes. Wildcard compares are basically useless (there is no prefix for immediates, for instance), so I have to fall through upon failure for each opcode assemble step, rather than error out like with other cores.
Not really a huge deal. My goal is to also write an x86 disassembler, and then combine the two for full x86 cross-platform detours, and that's only assembling 5-20 bytes :)
If it weren't so huge, I'd use it to assemble libco::x86's source code at startup. But bass is far larger than all of libco. Can at least use it for the reference block assembly going forward. But then I'd also need to do so for libco::amd64, and that's ... even more fun.