Update to bass v02r05 release.
Above-mentioned support [1] has been added, and I've also written up some basic documentation in the doc/ folder.
Nitpick away :P, but note that I don't intend to have some amazing book-length documentation. I expect people using this to know what they are doing.
I'd like to re-add something akin to xas' warnpc. The idea was that a warning/error would be thrown if the file write position had exceeded a certain value.
Say you were inserting a script into a ROM that only had 32KB of space for it, and your script was 33KB. This would serve as a warning that you've just crushed part of the ROM.
I'd like something more verbose this time, though, something like:
if condition
warning "Warning text!"
elseif otherCondition
error "Error text!"
endif
The obvious problem being that there is no way to express the current program counter to the condition clause. So we'd need something like a global-scope specialized define, like {pc}.
org 0xc08000
incbin script.bin
if {pc} >= 0xc10000
error "SIGTEDWOOLSEY: your script is too big, it just overwrote part of the game code!"
endif
Could also just define a label right before the if, but that's not as pretty.
If we do something like that, it'd make sense to simplify the print command to use it as well. But then I'm a bit afraid of crazy people using {pc} in opcodes, so it'd have to be done only for internal directives.
You see, it's up to the architecture to determine when to write data. snes-cpu will perform the evaluation before writing the opcode prefix, but x86 does not, due to handling of opcode prefix bytes for it.
Meh.
[1] Added nested if/elseif/else/endif support. It operates at the block-level, so they work inside of macros.
Note that this is for meta-programming / static compilation, eg conditional assemble, ala #ifdef in C.
It's not a cheap platform-specific mid-level assembly language directive.
Should be far superior to "rep 0/1" tricks. Will tell you if you forget an endif, or have too many of them.
Code:
	mapper lorom
	org $8000; fill $8000
	org $8000

	define size = -16

	if 0
	  if 0
	    lda #$01
	  elseif 0
	    lda #$02
	  else
	    lda #$03
	  endif
	elseif 1
	  if 0
	    lda #$04
	  elseif 1
	    lda #$05
	  else
	    lda #$06
	  endif
	else
	  if 0
	    lda #$07
	  elseif 0
	    lda #$08
	  else
	    lda #$09
	  endif
	endif
	  lda #$0a

	if {size} < 0
	  lda #$10
	elseif {size} > 0
	  lda #$20
	else
	  lda #$30
	endif
Added add (relative) seek command. It's like skip, but bi-directional.
Code:
	db $01,$02; seek -2; db $03,$04
Will write $03,$04
Code:
	db $01,$02; seek 2; db $03,$04
Will write $01,$02,{skip},{skip},$03,$04.
Positive seeking should only be done if the data is already filled in, obviously.
Otherwise, it'll probably end up as 0x00s.