Update to bass v02r08 release.
Updated the documentation. Defines are now argumentless macros instead of redefinable labels.
The reason for doing that is because labels are meant to be used before being defined, and thus cannot be redefined later.
By trying to shoehorn defines onto them for the nicer {}-less syntax, it allows the same undeclared define to go beyond the first pass.
Anyway, as a bonus for having to use {}, the evaluation happens post-expansion, allowing for the "lda #$00" -> byte-size detection to work, whereas lda #offset's size cannot be determined, as offset has already been evaluated to a raw integer.
Code:
	mapper lorom
	org $8000; fill $8000; org $8000

	macro add n; clc; adc {n}; endmacro
	macro sub n; sec; sbc {n}; endmacro
	macro addsub x,y
	  {add {x}}
	  {sub {y}}
	endmacro

	define offset $7efff8 + 8
	define length $20
	define 'A' 0x61
	define 'B' 0x62

	main:
	  {addsub #$20,#$40}
	  lda {offset}
	  {add #{length}}
	  db "ABCD"
	  jml main

	if {pc} >= 0x8040
	  error "Out of space! PC = {pc}"
	elseif {pc} >= 0x8020
	  warning "Almost out of space! PC = {pc}"
	endif

	macro m1 x
	  macro m2 y
	    {x} {y}
	  endmacro
	endmacro

	{m1 lda}; {m2 #$ff}

	macro test; nop #2; endmacro; {test}
The last idea I was considering was allowing more than one seek with +/- labels. Before with xas, I used to have +/-/++/--/+++/---. The syntax looked really ugly in code usage, however.
What I was thinking was that we define them with + and - only, but when evaluating them, allow ++ to mean (go to the second + forward) and -- to mean (go to the second - backward). Stupid example:
Code:
	-;
	  lda {bitpos}; tax
	  lda $2000,y
	  -; beq +; lsr; dex; bra -; +
	  sta $2000,y
	  dey; bne --
Agreed? How many should we reasonably allow? I was thinking anything more than three (+++/---) would be insane. I liked the +1 and -2 syntax, but those are also valid integer expressions for the math parser. I'd have to use a unique markup for them (like $ for hex and % for decimal) if we were to evaluate these inside of the math parser's tokenizer. But I don't see a lot of value in that. Only really obvious one would be something like:
Code:
	dw ++ - +; +; incbin data.bin; +
Which is probably better expressed anyway as:
Code:
	dw data.end - data; data: incbin data.bin; .end:
If we wanted the added flexibility, we'd have to use {+} and {-}.
Code:
	dw {++} - {+}; +; incbin data.bin; +
	dw {+2} - {+1}; +; incbin data.bin; +  //this could work too now
	-; lsr; dex; bne {-}
Is that worth it? Labels don't really use {} otherwise, so it's a bit unorthodox.