Update to bass v04r05 release.
Added a table for the SNES S-SMP (SPC700) processor. This uses my new-style syntax, but I'm making both the CPU+SMP use the "lda addr,x" style syntax instead of "lda addr+x" syntax.
As a result, for the S-SMP, I had to use = for argument separators, eg "and $08=$09" ($08 &= $09)
I really think that it should be "lda addr+x", but pretty much 100% of 6502 and 65816 programmers use "lda addr,x", so that's what we'll do.
SPC700 is rare, and I'm redefining the syntax either way, so this should be fine.
The table assembler syntax has proven to be quite capable now, but I do still have a few plans to extend it further.
It's also possible to inherit the table assembler and then add on to it for really tricky things (like ARM7 push/pop argument lists), if one wanted to do that (I'm not going to.)

[No archive available]