Update to bass v05r04 release.
At last, got it. You can now define a macro based on a condition, or a condition inside of a macro.
If you get stupid and try nesting them too deeply, I don't know if it will work or not. But the code is getting too complex for me to fix it if it doesn't work. My mind can only keep track of so many variables at one time.
Code:
	arch snes.cpu

	define styleA 1

	if {styleA} == 1
	macro read data
	db {data},{data}
	endmacro
	else
	macro read data
	db {data}
	endmacro
	endif

	macro write data
	if {data} == 1
	db {data}
	else
	db $bb
	endif
	endmacro

	org $0000; fill $8000
	org $0000
	if {styleA} == 1
	db $ff
	elseif {styleA} == 1
	db $ee
	endif

	{read $12}
	{write 1}

	//writes $ff,$12,$12,$01

[No archive available]