Update to bass v09r02 release.
This release adds tdc and tcd aliases for tda and tad.
It also reworks the expression evaluation. We now have evalBlock() to pull out a single block from one or more statements, and evalMacros() has been modified to work with this. activeLine was added to keep track of all the blocks on the current line, and evaluate them one at a time.
With these changes, macro expansion can occur inside the block parser, so we can now peek ahead and evaluate elseif {macro} statements when we are not inside a macro declaration and we are still capable of matching (eg not already matched) on the current condition level.
For right now, activeLine is a vector, meaning that inserting new blocks is an O(n) operation, meaning that recursive macro usage will expand at O(n^2). Not great. I need to write a linkedlist class, at which point I can insert at O(1), allowing for O(n) macro expansion, and we can then discard previously evaluated blocks to keep memory usage low. Anyway, it's only a problem if you expect to recurse inside macros more than a hundred thousand times or so.
If you guys can please test macros + conditions nested in weird patterns to look for bugs, that'd be greatly appreciated.
Once we confirm this latest change is working, I'll go ahead and release bass v10.