Update to bass v10r01 release.
So, the way macros expanded on the current line, and pushed the blocks out in-place, was quite arcane.
I've gutted macro support to improve things. Now, "{foo}" as a block is counted as a full-fledged macro, and its source is passed to recursive assembleSource() function (which assembleFile() also uses for incsrc recursion.) Code such as "lda {foo}" uses define mode, which will no longer expand to additional blocks, so you can't define foo as "#$55; nop" and expect that to work.
The compelling advantage of this is sane support for labels and macros inside of macros. The "self" namespace now exists within macros, which let you access macro arguments in isolation of global macros, and the same for labels. Example:
Code:
	macro clear value
	//you could say push namespace; namespace self; ...; pull namespace if you wanted to omit self::
	//but then global variables would need namespace prefixes, so ... up to you
	lda #{self::value}
	self::loop:
		stz $2180
		dec; bne self::loop
	endmacro
In this case, "{self::value}" evaluates to #{macroRecursionDepth}::value, and "self::loop" evaluates to #{macroExpansionCounter}::loop. In other words, you can call the clear macro more than once, or recurse on the clear macro more than once, and the names will no longer shadow the global namespace macros and labels.
With this, I've done away with {#}, as it's no longer necessary. I've also added {hex macro/label}, which is useful with print to display hexadecimal values.
...
Major caveat for this release is that conditional assembly is currently broken due to the macro changes. Going to be a bitch to get that working again :(

[No archive available]