Update to bass v11r01 release.
So this is an attempt to extend bass to be more than just a cross-assembler.
The idea being that we can use custom archs to perform DSL tasks. Not really sure what I have in mind just yet, though.
"-o output" is now an optional command-line argument. Attempting to write to a file that isn't opened will produce an error.
There is also the new "output "filename", mode" directive. "mode" is optional (it can be create or modify, defaults to the command-line setting: -overwrite ? create : modify)
So with this, you can use "bass file.asm", and have the ASM file specify which file to write to. Or if you want, which files.
I also added the "constant" directive, which creates labels. You can use them to create labels for functions inside a binary you are modifying, or you can use them for static, read-only computation. You cannot redefine them. They are to labels as defines are to macros, basically.
Code:
	lda addr,x  //lda $208000,x
	constant base $200000
	constant addr base + $8000
	if addr == $208000; ...; endif  //be careful, the first pass won't know the value of addr until it has been declared

[No archive available]