Update to bass v11r07 release.
Okay, this should be pretty much finished now. I'm out of ideas on useful things to add or change.
I'll certainly be very interested to see what kind of evil can be dreamed up with the new power.
I'd appreciate if people could stress test it, look for bugs, etc. If all looks good, I'll post this as v12, which should be stable for a while.