Update to bass v14r02 release.
This is the current WIP:
http://files.byuu.org/temp/bass_v14r02.tar.xz  (source only)
https://www.dropbox.com/s/mn7rqwm34jyk5 ... 02.7z?dl=0  (Windows 32-bit
binary)

This is the last stable release:
http://files.byuu.org/download/bass_v14.tar.xz  (source + Windows
binary)

I've updated bass to compile using the newest version of nall, but it's
quite a lot of changes so I'm worried that something may compile that
has changed semantics under the hood.

If possible, I'd appreciate if anyone using bass could try assembling
their projects with this new WIP, and then do a binary diff comparison
between what bass v14 and v14.02 produces, and let me know if they spot
any differences.

[No archive available]