Update to bass v14r07 release.
https://files.byuu.org/temp/bass_v14r07.tar.xz

I added SNES CPU "branch #immediate" instructions for qwertymodo, eg
`bra #$03`.

I added a new alternative macro, named inline. Simply replace the word
macro with inline, and you get a macro that has no stack frame and is
unscoped. Note that this is not the same as v14r05's unscoped macros.
This time, the stack frame doesn't store variables inside of it ... they
go into the existing stack frame you're currently in. Note that this may
not be either the global scope nor global stack frame. inline is meant
to represent a concept as if the code actually appeared right where you
included it, whereas macro is meant to indicate a functional style
programming that avoids polluting the current stack frame and can be
used for recursive invocations.

push and pull were renamed enqueue and dequeue, so as to avoid
conflicting with various ISAs that use push and/or pull as mnemonics.

putchar was renamed to print_char, and because it's so annoying to
create using said function, I also added print_hex. I also fixed a bug
where putchar before was not writing to the correct output handle, so
the buffers would get out of sync between it and print.

macro, define, variable lookup was radically changed in this release.
Instead of looking in the current stack frame, and then the global stack
frame, and then giving up ... it now searches recursively through all
stack frames, starting from the deepest (current) level, all the way up
to the top level. It tries elevating through the entire scope (eg
a.b.c.name, then a.b.name, then a.name, then name) before moving up to
the next stack frame and repeating. This probably won't actually be all
that useful in practice, but I'm sure you guys will come up with some
Eldritch horror that relies on it ;)

I restored the functionality of the global keyword to how it worked in
v14r05. It simply puts things in the global stack frame ... it doesn't
remove scoping anymore.

I sincerely hope you guys like this WIP. I tried to address all your
concerns with r06, and it took a hell of a lot of work >_>