Update to bass v14r08 release.
https://files.byuu.org/temp/bass_v14r08.tar.xz

See above post for changes.

The last thing to solve is the scope escaping stuff. Once we're happy
with that, it'll be time to push v15.

------------------------------------------------------------------------


Side note: my goal is strict backward compatibility with v15 in the
future.

I'm okay with bass forks, but would the existing bass forks consider
rebasing their own work off v15 in the future? Doesn't have to be right
away.

I was thinking, if every bass variant has the same core language syntax,
then it will become possible to produce bass code that works in every
fork.

Obviously, it's expected that each fork will add additional features.
And code that uses those features will lose compatibility with other
bass forks. But so long as they don't modify the core language syntax,
this should be acceptable.

I can't force peoples' hands here, so if it's a situation of fork
maintainers not wanting to rewrite existing code to the new syntax, I'll
understand that. I can't even 100% guarantee I won't end up breaking v15
code again in the future, but will try really hard not to. Still, it
would be a shame to finally try to aim for stability but still have
forks that limit this idea.